Move tests to integration_test
