+++ packages/google_sign_in/google_sign_in_web/test/pubspec.yaml
@@ -1,0 +1,23 @@
+name: regular_integration_tests
+publish_to: none
+
+environment:
+  sdk: ">=2.2.2 <3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+
+dev_dependencies:
+  google_sign_in: ^4.5.3
+  flutter_driver:
+    sdk: flutter
+  flutter_test:
+    sdk: flutter
+  http: ^0.12.2
+  mockito: ^4.1.1
+  google_sign_in_web:
+    path: ../
+  integration_test:
+    path: ../../../integration_test
+
