+++ packages/google_sign_in/google_sign_in_web/test/run_test
@@ -1,0 +1,17 @@
+#!/usr/bin/bash
+if pgrep -lf chromedriver > /dev/null; then
+  echo "chromedriver is running."
+
+  if [ $# -eq 0 ]; then
+    echo "No target specified, running all tests..."
+    find test_driver/ -iname *_integration.dart | xargs -n1 -i -t flutter drive -d web-server --web-port=7357 --browser-name=chrome --target='{}'
+  else
+    echo "Running test target: $1..."
+    set -x
+    flutter drive -d web-server --web-port=7357 --browser-name=chrome --target=$1
+  fi
+
+  else
+    echo "chromedriver is not running."
+fi
+
