+++ packages/google_sign_in/google_sign_in_web/test/test_driver/auth2_integration.dart
@@ -8,7 +8,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:google_sign_in_platform_interface/google_sign_in_platform_interface.dart';
 import 'package:google_sign_in_web/google_sign_in_web.dart';
 import 'gapi_mocks/gapi_mocks.dart' as gapi_mocks;
-import 'utils.dart';
+import 'src/test_utils.dart';
 
 void main() {
   GoogleSignInTokenData expectedTokenData =
