+++ packages/google_sign_in/google_sign_in_web/test/test_driver/gapi_load_integration_test.dart
@@ -1,0 +1,8 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:integration_test/integration_test_driver.dart';
+
+Future<void> main() async => integrationDriver();
+
